Moved getDataFromFile back to main, still doesn't work

+++ Normal.cpp
@@ -22,7 +22,7 @@ Normal::Normal() : Player(){
 }
 
 Normal::Normal(int rank, string name, int born, string position, int gamesPlayed, int goals,
-               int assists, int points, int penalty, int fightStat, int shotStat, int passStat) :
+               int assists, int points, int penalty, double fightStat, double shotStat, double passStat) :
                Player(rank, name, born, position, gamesPlayed,goals, assists,
                       points, penalty, fightStat, shotStat, passStat){
 }
