+++ Normal.h
@@ -12,7 +12,7 @@ public:
     Normal();
 
     explicit Normal(int rank, string name, int born, string position, int gamesPlayed,
-                    int goals, int assists, int points, int penalty, int fightStat, int shotStat, int passStat);
+                    int goals, int assists, int points, int penalty, double fightStat, double shotStat, double passStat);
 
 
     virtual std::string playerSaying() override;
